Fixed seed to support SQLite

+++ dev/db-reset.sh
@@ -1,5 +1,17 @@
 #!/bin/sh
-dropdb -h localhost -U postgres bridg
-createdb -h localhost -U postgres bridg
+
+# FIXME: Replace with Python script to handle DATABASE_URI
+
+# Reset PostgreSQL
+if pg_isready -h localhost >/dev/null; then
+  dropdb -h localhost -U postgres bridg
+  createdb -h localhost -U postgres bridg
+else
+  echo "NOTE: PostgreSQL is not running"
+fi
+
+# Reset SQLite
+rm -f tmp/bridg.db
+
 python -m dev.migrate
 python -m dev.seed
